whisperfy-backend/requirements.txt
fastapi
uvicorn
openai-whisper
requests
pydantic
python-multipart
stripe
python-dotenv
firebase-admin